[WFCORE-6813]:  Improving testing
